viewの作成
